requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
dnspython==1.16.0
dotenv-cli==3.2.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
frozenlist==1.4.1
gnews==0.3.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.6
jieba3k==0.35.1
joblib==1.3.2
lxml==5.2.0
lxml_html_clean==0.1.0
multidict==6.0.5
newspaper3k==0.2.8
nltk==3.8.1
openai==1.14.3
pillow==10.2.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
